Convert the fd scan limit through int
rlim_t is a signed long on FreeBSD, so casting it to long is redundant
there and GCC 16 reports -Wuseless-cast; on Linux and macOS rlim_t is an
unsigned 64-bit type and the cast is required. The guarding branch has
already proved the value is no greater than INT_MAX, so int is both safe
and never the same type as rlim_t on any supported platform.

+++ src/env.c
@@ -2287,7 +2287,14 @@ static long p101_env_exec_scan_limit(void)
         }
         else
         {
-            limit = (long)rl.rlim_cur;
+            /*
+             * rlim_t is a signed long on FreeBSD and an unsigned 64-bit type
+             * elsewhere, so a cast to long is redundant on one platform and
+             * required on the others. The branch above has already proved the
+             * value is no greater than INT_MAX, so convert through int, which
+             * is never the same type as rlim_t on any supported platform.
+             */
+            limit = (int)rl.rlim_cur;
         }
     }
